requirements.txt
# -----------------------------------------------------------------------------
# Python Dependencies for the Automated Video Pipeline
# -----------------------------------------------------------------------------
# Install all these libraries using the command:
# pip install -r requirements.txt
#
# For PyTorch with GPU support, see the special command in setup.sh.
# -----------------------------------------------------------------------------

# --- Core Libraries ---
google-generativeai # For all Gemini API interactions
pydub             # For audio manipulation and silence detection
Pillow            # For image processing (used in tests and potentially resizing)

# --- PyTorch (for Whisper Model) ---
# Whisper runs on PyTorch. These are the core components.
# For GPU (CUDA) acceleration, these must be installed with a specific index URL.
torch
torchaudio
torchvision

# --- Whisper Model Library (Installed from GitHub) ---
# This specific fork provides the essential word-level timestamping.
# pip will install this directly from the git repository.
git+https://github.com/linto-ai/whisper-timestamped.git